made things floats because ¯\_(ツ)_/¯

+++ pythonDnd/pythag.py
@@ -1,7 +1,7 @@
 import math
 x = float(input("\nx to creature: "))
 y = float(input("y to creature: "))
-spellRange = int(input("spell/attack range: "))
+spellRange = float(input("spell/attack range: "))
 
 def rangeFinder(xToTarg,yToTarg,rangeOfAttack):
     hyp = math.sqrt((x**2)+(y**2))
